requirements.txt
anthropic>=0.18.0  # remove openai package
langchain>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
logging>=0.5.1.2
black>=23.9.1
isort>=5.12.0
mypy>=1.5.1
jupyter>=1.0.0
notebook>=7.0.0